[#3] - create CRUD repository methods and post service with validations

+++ src/main/resources/db/posts/changelog/changes/001.sql
@@ -1,7 +1,7 @@
 create table posts
 (
     id                      uuid primary key,
-    account_id              integer not null,
+    account_id              bigint not null,
     content                 varchar(200),
     version                 integer not null,
     created_date_timestamp  timestamp default current_timestamp,
